+++ src/codec/csv.ts
@@ -1,14 +1,12 @@
-import { z } from "zod";
+import { stringSplit } from "./string-split.js";
 
-const codec = z.codec(z.string(), z.array(z.string()), {
-  decode: (str) => str.split(","),
-  encode: (arr) => arr.join(","),
-});
+const codec = stringSplit(",");
 
 /**
  * Returns a codec that takes in a CSV string and outputs a string array.
  *
- * The string is splitted on the "," character.
+ * The string is splitted on the "," character. It uses `stringSplit(",")`
+ * internally.
  *
  * @returns A CSV codec.
  */