+++ src/codec/index.ts
@@ -1,13 +1,13 @@
 import { csv } from "./csv.js";
 import { jsonParse } from "./json-parse.js";
-import { stringtoInteger } from "./string-to-integer.js";
+import { stringToInteger } from "./string-to-integer.js";
 import { stringToJson } from "./string-to-json.js";
 import { stringToUrl } from "./string-to-url.js";
 
 const codec = {
   csv,
   jsonParse,
-  stringtoInteger,
+  stringtoInteger: stringToInteger,
   stringToJson,
   stringToUrl,
 };