+++ CHANGELOG.md
@@ -10,10 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `zu.string.json()` with branded type `JsonString`.
+- `zu.codec.jsonParse(schema)`, a codec factory that first transforms the string input
+using JSON.parse, then forwards the validation to the provided schema.
 
 ### Changed
 
-- Moved `zu.json.stringified()` to `zu.codec.stringToJson()`, and made it a codec!
+- Moved `zu.json.stringified()` to `zu.codec.stringToJson()`, and made it a codec! The
+codec uses `zu.codec.jsonParse(json())` internally.
 
 ## [0.20.0] - 2026-01-20